fixed organization of endpoints so it is easy to follow
fixed endpoint organization
