+++ moosement_backend/flask-server/app.py
@@ -1,0 +1,28 @@
+# app.py
+from flask import Flask
+from flask_cors import CORS
+from challenges import challenges_bp  # Import the challenges blueprint
+from users import users_bp  # Example of another blueprint for user-related routes
+from teams import teams_bp  # Example of another blueprint for team-related routes
+
+def create_app():
+    """
+    Initializes and configures the Flask application.
+    Registers all necessary blueprints.
+    """
+    app = Flask(__name__)
+    CORS(app)  # Enable CORS for frontend communication
+
+    # Register the Blueprint for challenges
+    app.register_blueprint(challenges_bp, url_prefix='/api/challenges')
+    
+    # Register the Blueprint for users (if you have one)
+    app.register_blueprint(users_bp, url_prefix='/api/users')
+
+    # Register the Blueprint for teams (if you have one)
+    app.register_blueprint(teams_bp, url_prefix='/api/teams')
+
+    # Register any additional blueprints here
+    # e.g., app.register_blueprint(other_bp, url_prefix='/api/other')
+
+    return app
