+++ moosement_backend/flask-server/insert_challenges.py
